requirements/auto_generated_requirements.txt

amqp==2.4.2
billiard==3.6.0.0
bitmapist==3.97
boto3==1.9.77
boto==2.49.0
botocore==1.12.133
celery[redis]==4.3.0
certifi==2019.3.9
chardet==3.0.4
clom==0.8.0a1
django-absolute==0.3
django-ace==1.0.5
django-appconf==1.0.3
django-bootstrap3==11.0.0
django-bower==5.1.0
django-cacheops==4.1
django-dynamic-preferences==1.7.0
django-extensions==2.1.6
django-filter==2.1.0
django-nvd3==0.9.7
django-redis==4.10.0
django-select2==6.3.1
django-ses==0.8.2
django-storages==1.6.5
django-suit==0.2.26
django-tables2==2.0.6
django-tinymce==2.6.1
django-versionfield3==0.1.8
django==2.2
djangorestframework==3.7.7
docutils==0.14
ecdsa==0.13
funcsigs==1.0.2
funcy==1.12
furl==0.4.93
future==0.16.0
google-cloud==0.34.0
idna==2.8
ipaddress==1.0.22
jinja2==2.10.1
jmespath==0.9.4
jsonfield==1.0.3
kombu==4.5.0
markdown==2.6.11
markupsafe==1.1.1
msgpack-python==0.4.7
orderedmultidict==1.0
paver==1.2.4
persisting-theory==0.2.1
pillow==6.0.0
protobuf3-to-dict==0.1.5
protobuf==3.7.1
psycopg2==2.7.3.2
python-dateutil==2.8.0 ; python_version >= '2.7'
python-magic==0.4.15
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2016.4
raven==6.10.0
redis==3.2.1
requests==2.21.0
s3transfer==0.1.13
singledispatch==3.4.0.3
six==1.10.0
sqlparse==0.3.0
unidecode==1.0.23
urllib3==1.24.2 ; python_version >= '3.4'
uwsgi==2.0.18
vine==1.3.0
asn1crypto==0.24.0
aws-sam-translator==1.11.0
aws-xray-sdk==2.4.2
blessings==1.7
boto3==1.9.77
boto==2.49.0
botocore==1.12.133
bumpversion==0.5.3
certifi==2019.3.9
cffi==1.12.3
cfn-lint==0.19.1
chardet==3.0.4
coverage==4.0.3
cryptography==2.6.1
cssselect==1.0.3
django-debug-panel==0.8.3
django-debug-toolbar==1.11
django-nose==1.4.6
django==2.2
docker-pycreds==0.4.0
docker==3.7.2
docutils==0.14
ebs-deploy==2.0.1
ecdsa==0.13
factory-boy==2.8.1
faker==1.0.5
freezegun==0.3.7
future==0.16.0
idna==2.8
jinja2==2.10.1
jmespath==0.9.4
jsondiff==1.1.2
jsonpatch==1.23
jsonpickle==1.1
jsonpointer==2.0
jsonschema==2.6.0
lxml==4.3.3
markupsafe==1.1.1
mock==2.0.0
moto==1.3.8
nose-progressive==1.5.1
nose==1.3.7
pbr==5.1.3
pluggy==0.9.0
py==1.8.0
pyasn1==0.4.5
pycparser==2.19
pyquery==1.2.13
python-coveralls==2.9.1
python-dateutil==2.8.0 ; python_version >= '2.7'
python-jose==3.0.1
pytz==2016.4
pyyaml==5.1
requests==2.21.0
responses==0.10.6
rsa==4.0
s3transfer==0.1.13
six==1.10.0
sqlparse==0.3.0
text-unidecode==1.2
tox==2.5.0
urllib3==1.24.2 ; python_version >= '3.4'
virtualenv==16.5.0
websocket-client==0.56.0
werkzeug==0.15.2
wrapt==1.11.1
xmltodict==0.12.0
xmlunittest==0.5.0
